refactor into keyring class
